bulk_sensor: Move APIDumpHelper() from motion_report.py to bulk_sensor.py
