+++ klippy/extras/angle.py
@@ -4,7 +4,7 @@
 #
 # This file may be distributed under the terms of the GNU GPLv3 license.
 import logging, math
-from . import bus, motion_report, bulk_sensor
+from . import bus, bulk_sensor
 
 MIN_MSG_TIME = 0.100
 TCODE_ERROR = 0xff
@@ -441,7 +441,7 @@ class Angle:
         mcu.register_config_callback(self._build_config)
         self.bulk_queue = bulk_sensor.BulkDataQueue(mcu, "spi_angle_data", oid)
         # API server endpoints
-        self.api_dump = motion_report.APIDumpHelper(
+        self.api_dump = bulk_sensor.APIDumpHelper(
             self.printer, self._api_update, self._api_startstop, 0.100)
         self.name = config.get_name().split()[1]
         wh = self.printer.lookup_object('webhooks')
