+++ klippy/extras/mpu9250.py
@@ -5,7 +5,7 @@
 #
 # This file may be distributed under the terms of the GNU GPLv3 license.
 import logging, time
-from . import bus, motion_report, adxl345, bulk_sensor
+from . import bus, adxl345, bulk_sensor
 
 MPU9250_ADDR =      0x68
 
@@ -80,7 +80,7 @@ class MPU9250:
                                                           BYTES_PER_SAMPLE)
         self.last_error_count = 0
         # API server endpoints
-        self.api_dump = motion_report.APIDumpHelper(
+        self.api_dump = bulk_sensor.APIDumpHelper(
             self.printer, self._api_update, self._api_startstop, API_UPDATES)
         self.name = config.get_name().split()[-1]
         wh = self.printer.lookup_object('webhooks')
